make changes to cleaning
